organized code, updated readme, added changelog and code of conduct
