+++ .gitignore
@@ -1,1 +1,2 @@
 vendor
+.DS_Store
