+++ CHANGELOG.md
@@ -1,0 +1,30 @@
+# Version 0.3.4
+
+* Updated README
+* Added CHANGELOG
+* Added Contributor Code of Conduct
+* Organized Code
+
+# Version 0.3.3
+
+* Hotfix autoload path
+
+# Version 0.3.2
+
+* Added Show Admin Url Command, added `bin` to composer.json
+
+# Version 0.3.1
+
+* Change readme: Installation via Packagist
+
+# Version 0.3.0
+
+* `NewCommand` now installs PW instead of just downloading it (thanks to @HariKT), added Commands for Fields, Modules, Backup
+
+# Version 0.2.0
+
+* Added Create Template Command, extended Create User Command
+
+# Version 0.1.0
+
+* Initial
