+++ THANKS.md
@@ -1,0 +1,3 @@
+# THANKS
+
+[HariKT](https://github.com/harikt) for his big contribution of a real "NewCommand" and command line based installer within Wireshell!
