+++ src/Commands/Common/NewCommand.php
@@ -9,7 +9,7 @@
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
-namespace Wireshell\Commands;
+namespace Wireshell\Commands\Common;
 
 use Distill\Distill;
 use Distill\Exception\IO\Input\FileCorruptedException;
